Change: note length is represented in uint32_t

+++ firmware/src/speaker.hpp
@@ -33,13 +33,13 @@ namespace gb7::sound
         struct Note
         {
             Tone tone;
-            uint64_t length; // microseconds
+            uint32_t length; // microseconds
         };
 
         static inline queue<Note, 128> m_notes;
         static inline Tone m_tone = Tone::None;
-        static inline uint64_t m_count = 0;
-        static inline uint64_t m_count_to = 0;
+        static inline uint32_t m_count = 0;
+        static inline uint32_t m_count_to = 0;
 
     public:
         inline void init()
